SVG file for the emoji
Used in two ways--to print the graphic (glued to cardboard backing) and to generate hole pattern for plate.scad (holes for eyes)
recommendation: when printing, to fit the printer the graphic was scaled down. this meant it did not match the size. by deleting the eyes before printing (printing just the big yellow circle with the mouth) it did not matter as much. just poke a hole in the cardboard big enough for the motor shaft.